feat: create Sign Up Copy (#3880)
* feat: create Sign Up Copy

* fix: style fix

* fix: reinstate success dto

* fix: update icon margin

---------
