+++ sites/public/src/pages/sign-in.tsx
@@ -15,11 +15,13 @@ import {
 } from "@bloom-housing/shared-helpers"
 import { UserStatus } from "../lib/constants"
 import { EnumUserErrorExtraModelUserErrorMessages } from "@bloom-housing/backend-core/types"
-import { HeadingGroup } from "@bloom-housing/ui-seeds"
 import SignUpBenefits from "../components/account/SignUpBenefits"
+import signUpBenefitsStyles from "../../styles/sign-up-benefits.module.scss"
+import SignUpBenefitsHeadingGroup from "../components/account/SignUpBenefitsHeadingGroup"
 
 const SignIn = () => {
   const { login, userService } = useContext(AuthContext)
+  const signUpCopy = process.env.showMandatedAccounts
   /* Form Handler */
   // This is causing a linting issue with unbound-method, see open issue as of 10/21/2020:
   // https://github.com/react-hook-form/react-hook-form/issues/2887
@@ -128,69 +130,46 @@ const SignIn = () => {
     }
   }, [networkError])
 
-  const SignUpBenefitsHeadingGroup = (props: { mobileView: boolean }) => {
-    const classNames = props.mobileView ? "py-6 px-4" : ""
-    return (
-      <HeadingGroup
-        heading={t("account.signUpSaveTime.title")}
-        subheading={t("account.signUpSaveTime.subTitle")}
-        size="2xl"
-        className={classNames}
-      />
-    )
-  }
-
   return (
     <>
-      {process.env.showMandatedAccounts ? (
-        <FormsLayout className="sm:max-w-lg md:max-w-full">
-          <div className="flex flex-col md:flex-row md:ml-20 justify-center">
-            <div className="display md:hidden ">
+      <FormsLayout className={signUpCopy && "sm:max-w-lg md:max-w-full"}>
+        <div className={signUpCopy && signUpBenefitsStyles["benefits-container"]}>
+          {signUpCopy && (
+            <div className={signUpBenefitsStyles["benefits-display-hide"]}>
               <SignUpBenefitsHeadingGroup mobileView={true} />
             </div>
-            <div className="md:max-w-lg w-full justify-center">
-              <FormSignIn
-                onSubmit={(data) => void onSubmit(data)}
-                control={{ register, errors, handleSubmit, watch }}
-                networkStatus={{
-                  content: networkStatusContent,
-                  type: networkStatusType,
-                  reset: () => {
-                    reset()
-                    resetNetworkError()
-                    setConfirmationStatusMessage(undefined)
-                  },
-                }}
-                showRegisterBtn={true}
-              />
-            </div>
-            <div className="hidden md:flex">
-              <div className="md:flex md:flex-col md:p-8 md:max-w-lg md:w-full ">
+          )}
+          <div className={signUpCopy && signUpBenefitsStyles["benefits-form-container"]}>
+            <FormSignIn
+              onSubmit={(data) => void onSubmit(data)}
+              control={{ register, errors, handleSubmit, watch }}
+              networkStatus={{
+                content: networkStatusContent,
+                type: networkStatusType,
+                reset: () => {
+                  reset()
+                  resetNetworkError()
+                  setConfirmationStatusMessage(undefined)
+                },
+              }}
+              showRegisterBtn={true}
+            />
+          </div>
+          {signUpCopy && (
+            <div className={signUpBenefitsStyles["benefits-hide-display"]}>
+              <div className={signUpBenefitsStyles["benefits-desktop-container"]}>
                 <SignUpBenefitsHeadingGroup mobileView={false} />
                 <SignUpBenefits idTag="desktop" />
               </div>
             </div>
-            <SignUpBenefits idTag="mobile" className="display md:hidden" />
-          </div>
-        </FormsLayout>
-      ) : (
-        <FormsLayout>
-          <FormSignIn
-            onSubmit={(data) => void onSubmit(data)}
-            control={{ register, errors, handleSubmit, watch }}
-            networkStatus={{
-              content: networkStatusContent,
-              type: networkStatusType,
-              reset: () => {
-                reset()
-                resetNetworkError()
-                setConfirmationStatusMessage(undefined)
-              },
-            }}
-            showRegisterBtn={true}
-          />
-        </FormsLayout>
-      )}
+          )}
+          {signUpCopy && (
+            <div className={signUpBenefitsStyles["benefits-display-hide"]}>
+              <SignUpBenefits idTag="mobile" />
+            </div>
+          )}
+        </div>
+      </FormsLayout>
 
       <ResendConfirmationModal
         isOpen={confirmationStatusModal}
