+++ sites/public/styles/sign-up-benefits.module.scss
@@ -1,0 +1,54 @@
+.benefits-form-layout {
+    @media (min-width: theme("screens.sm")) {
+      max-width: var(--seeds-width-lg);
+    }
+    @media (min-width: theme("screens.md")) {
+      max-width: 100%;
+    }
+  }
+  .benefits-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+  
+    @media (min-width: theme("screens.md")) {
+      flex-direction: row;
+      margin-left: var(--seeds-s20);
+    }
+  }
+  
+  .benefits-display-hide {
+    display: block;
+  
+    @media (min-width: theme("screens.md")) {
+      display: none;
+    }
+  }
+  
+  .benefits-hide-display {
+    display: none;
+  
+    @media (min-width: theme("screens.md")) {
+      display: flex;
+    }
+  }
+  
+  .benefits-form-container {
+    width: 100%;
+    justify-content: center;
+  
+    @media (min-width: theme("screens.md")) {
+      max-width: var(--seeds-width-lg);
+      width: max-content;
+    }
+  }
+  
+  .benefits-desktop-container {
+    @media (min-width: theme("screens.md")) {
+      display: flex;
+      flex-direction: column;
+      padding: var(--seeds-s8);
+      max-width: var(--seeds-width-lg);
+      width: 100%;
+    }
+  }
